requirements.txt
django==3.2.5
djangorestframework==3.12.4
boto3==1.17.103
django-storages==1.11.1
